requirements.txt
discord.py
yt-dlp
google-genai
aiohttp